Change doubleclick setting to an int.

+++ glrender.cc
@@ -839,8 +839,7 @@ void mouse(int button, int state, int x, int y)
     glutMotionFunc(NULL);
     glutAttachMenu(button);
     Menu=true;
-    glutTimerFunc((unsigned int)
-                  (getSetting<double>("doubleclick")),timeout,0);
+    glutTimerFunc(getSetting<double>("doubleclick"),timeout,0);
     return;
   }
